+++ src/Neo.Extensions/RandomExtensions.cs
@@ -1,0 +1,34 @@
+// Copyright (C) 2015-2024 The Neo Project.
+//
+// RandomExtensions.cs file belongs to the neo project and is free
+// software distributed under the MIT software license, see the
+// accompanying file LICENSE in the main directory of the
+// repository or http://www.opensource.org/licenses/mit-license.php
+// for more details.
+//
+// Redistribution and use in source and binary forms with or without
+// modifications are permitted.
+
+using System;
+using System.Numerics;
+
+namespace Neo.Extensions
+{
+    public static class RandomExtensions
+    {
+        public static BigInteger NextBigInteger(this Random rand, int sizeInBits)
+        {
+            if (sizeInBits < 0)
+                throw new ArgumentException("sizeInBits must be non-negative");
+            if (sizeInBits == 0)
+                return 0;
+            Span<byte> b = stackalloc byte[sizeInBits / 8 + 1];
+            rand.NextBytes(b);
+            if (sizeInBits % 8 == 0)
+                b[^1] = 0;
+            else
+                b[^1] &= (byte)((1 << sizeInBits % 8) - 1);
+            return new BigInteger(b);
+        }
+    }
+}
