+++ src/Neo/Cryptography/ECC/ECCurve.cs
@@ -9,6 +9,7 @@
 // Redistribution and use in source and binary forms with or without
 // modifications are permitted.
 
+using Neo.Extensions;
 using System.Globalization;
 using System.Numerics;
