+++ tests/Neo.Cryptography.MPTTrie.Tests/Cryptography/MPTTrie/UT_Trie.cs
@@ -17,7 +17,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
-using static Neo.Helper;
+using static Neo.Extensions.HashSetExtensions;
 
 namespace Neo.Cryptography.MPTTrie.Tests
 {
