+++ tests/Neo.Network.RPC.Tests/UT_Nep17API.cs
@@ -11,6 +11,7 @@
 
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Moq;
+using Neo.Extensions;
 using Neo.Json;
 using Neo.SmartContract;
 using Neo.SmartContract.Native;
@@ -19,7 +20,7 @@ using Neo.Wallets;
 using System.Linq;
 using System.Numerics;
 using System.Threading.Tasks;
-using static Neo.Helper;
+using static Neo.Extensions.HashSetExtensions;
 
 namespace Neo.Network.RPC.Tests
 {
