+++ tests/Neo.UnitTests/Cryptography/ECC/UT_ECFieldElement.cs
@@ -12,6 +12,7 @@
 using FluentAssertions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Neo.Cryptography.ECC;
+using Neo.Extensions;
 using System;
 using System.Globalization;
 using System.Numerics;
