+++ tests/Neo.UnitTests/Cryptography/ECC/UT_ECPoint.cs
@@ -12,6 +12,7 @@
 using FluentAssertions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Neo.Cryptography.ECC;
+using Neo.Extensions;
 using Neo.IO;
 using System;
 using System.IO;
