+++ tests/Neo.UnitTests/Cryptography/UT_Base58.cs
@@ -12,6 +12,7 @@
 using FluentAssertions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Neo.Cryptography;
+using Neo.Extensions;
 using System;
 using System.Collections.Generic;
