+++ tests/Neo.UnitTests/Cryptography/UT_Crypto.cs
@@ -12,6 +12,7 @@
 using FluentAssertions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Neo.Cryptography;
+using Neo.Extensions;
 using Neo.Wallets;
 using System;
 using System.Security.Cryptography;
