+++ tests/Neo.UnitTests/IO/UT_MemoryReader.cs
@@ -10,6 +10,7 @@
 // modifications are permitted.
 
 using Microsoft.VisualStudio.TestTools.UnitTesting;
+using Neo.Extensions;
 using Neo.IO;
 using System.IO;
 using System.Text;
