+++ tests/Neo.UnitTests/Wallets/NEP6/UT_NEP6Contract.cs
@@ -11,6 +11,7 @@
 
 using FluentAssertions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
+using Neo.Extensions;
 using Neo.Json;
 using Neo.SmartContract;
 using Neo.Wallets.NEP6;
